+++ python/mmwcore/io/__init__.py
@@ -15,6 +15,11 @@ from .dca1000_types import (
     DCA1000CLITimeoutError,
     DCA1000NetworkConfig,
 )
+from .mmwcli_capture import (
+    MMWCLI_CAPTURE_SESSION_SCHEMA_V1,
+    ADCFileCapture,
+    open_capture,
+)
 from .packets import (
     DCA1000_PACKET_HEADER_BYTES,
     DCA1000_PACKET_PAYLOAD_BYTES,
@@ -32,6 +37,7 @@ from .serial import RadarSerialConfig, RadarSerialController
 __all__ = [
     "DCA1000_PACKET_HEADER_BYTES",
     "ADCFileFrameReader",
+    "ADCFileCapture",
     "DCA1000_PACKET_PAYLOAD_BYTES",
     "DCA1000_PACKET_PAYLOAD_INT16_VALUES",
     "DCA1000CLIController",
@@ -44,12 +50,14 @@ __all__ = [
     "DCA1000PacketSource",
     "DCA1000Packet",
     "PacketLossStats",
+    "MMWCLI_CAPTURE_SESSION_SCHEMA_V1",
     "RadarSerialConfig",
     "RadarSerialController",
     "assemble_dca1000_frame",
     "assemble_dca1000_frame_bytes",
     "load_adc_cube",
     "load_adc_file",
+    "open_capture",
     "parse_dca1000_packet",
     "read_dca1000_frame_from_packets",
     "reorder_dca1000_packets",